\name{bgmyc.dataprep}
\alias{bgmyc.dataprep}
\title{Prepares a tree for analysis.}
\description{
Takes an ape "phylo" object and turns it into input for downstream functions.
}
\usage{
bgmyc.dataprep(tr)
}

\arguments{
  \item{tr}{
an ape "phylo" object. 
}
}
\author{
Noah M. Reid
Maintainer: Dmitry Karabanov <dk@ibiw.ru>
}
