\name{plot.bgmycrates}
\alias{plot.bgmycrates}
\title{
Plots relative coalescent and Yule rates for model fit evaluation.
}
\description{
Plots relative coalescent and Yule rates for model fit evaluation.
}
\usage{
\method{plot}{bgmycrates}(x, ...)
}
\arguments{
\item{x}{
output from checkrates function.}
\item{...}{
additional arguments passed to plot.}
}
\details{
plots relative coalescent and Yule rates for model fit evaluation}
\author{
Noah M. Reid
Maintainer: Dmitry Karabanov <dk@ibiw.ru>
}
